fixed upload/replace.

+++ app/Http/Controllers/PostsController.php
@@ -151,13 +151,7 @@ class PostsController extends Controller
             'body' => 'required', 
             'cover_image' => 'image|nullable|mimes:jpeg,png,jpg,gif,svg|max:1999'
         ]);
-        
-        // Create Post
-        $post = Post::find($id);
-        $post->title = $request->input('title');
-        $post->body = $request->input('body');
-        $post->save();
-        
+
         // File upload handling 
         /**
          * Delete the old image if the image was updated.
@@ -175,18 +169,24 @@ class PostsController extends Controller
             $extension = $request->file('cover_image')->getClientOriginalExtension();
             // Filename to store
             $fileNameToStore= $filename.'_'.time().'.'.$extension;
-
-            if ($post->cover_image != 'noimage.jpg') { 
-                Storage::delete('public/cover_images/'.$post->cover_image); 
-            } 
-            $post->cover_image = $fileNameToStore;
             // Upload Image
             $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
             
         } 
-        else {
-            $fileNameToStore = 'noimage.jpg';
+        
+        // Create Post
+        $post = Post::find($id);
+        $post->title = $request->input('title');
+        $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+                if ($post->cover_image != 'noimage.jpg') {
+                    Storage::delete('public/cover_images/'.$post->cover_image);
+                }
+                $post->cover_image = $fileNameToStore;
         }
+        $post->save();
+        
+        
             
 
         //return redirect('/posts')->with('success', 'Post Created');
@@ -218,6 +218,11 @@ class PostsController extends Controller
         // Check for correct user.
         if(auth()->user()->id != $post->user_id){
             return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
         }
         return redirect('/posts')->with('success', 'Post Removed');
     }
